Refactoring: change ScrollTopButton to typescript
